Create SideMapControlOverlays component
